drivers device naming consistency
