+++ src/drivers/device/posix/SPI.hpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (C) 2015 Mark Charlebois. All rights reserved.
+ *   Copyright (C) 2017 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -30,10 +30,25 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  ****************************************************************************/
-#pragma once
 
-#ifdef __PX4_NUTTX
-#include "nuttx/I2C.hpp"
-#else
-#include "posix/I2C.hpp"
-#endif
+/**
+ * @file spi.h
+ *
+ * Base class for devices connected via SPI.
+ */
+
+#ifndef _DEVICE_SPI_H
+#define _DEVICE_SPI_H
+
+#include "CDev.hpp"
+
+#include <px4_spi.h>
+
+namespace device __EXPORT
+{
+
+// TODO: implement posix spi
+
+} // namespace device
+
+#endif /* _DEVICE_SPI_H */
