+++ src/drivers/device/spi.h
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (C) 2012 PX4 Development Team. All rights reserved.
+ *   Copyright (C) 2017 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -30,124 +30,10 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  ****************************************************************************/
+#pragma once
 
-/**
- * @file spi.h
- *
- * Base class for devices connected via SPI.
- */
-
-#ifndef _DEVICE_SPI_H
-#define _DEVICE_SPI_H
-
-#include "device.h"
-
-#include <px4_spi.h>
-
-namespace device __EXPORT
-{
-
-/**
- * Abstract class for character device on SPI
- */
-class __EXPORT SPI : public CDev
-{
-protected:
-	/**
-	 * Constructor
-	 *
-	 * @param name		Driver name
-	 * @param devname	Device node name
-	 * @param bus		SPI bus on which the device lives
-	 * @param device	Device handle (used by SPI_SELECT)
-	 * @param mode		SPI clock/data mode
-	 * @param frequency	SPI clock frequency
-	 */
-	SPI(const char *name,
-	    const char *devname,
-	    int bus,
-	    uint32_t device,
-	    enum spi_mode_e mode,
-	    uint32_t frequency);
-	virtual ~SPI();
-
-	/**
-	 * Locking modes supported by the driver.
-	 */
-	enum LockMode {
-		LOCK_PREEMPTION,	/**< the default; lock against all forms of preemption. */
-		LOCK_THREADS,		/**< lock only against other threads, using SPI_LOCK */
-		LOCK_NONE		/**< perform no locking, only safe if the bus is entirely private */
-	};
-
-	virtual int	init();
-
-	/**
-	 * Check for the presence of the device on the bus.
-	 */
-	virtual int	probe();
-
-	/**
-	 * Perform a SPI transfer.
-	 *
-	 * If called from interrupt context, this interface does not lock
-	 * the bus and may interfere with non-interrupt-context callers.
-	 *
-	 * Clients in a mixed interrupt/non-interrupt configuration must
-	 * ensure appropriate interlocking.
-	 *
-	 * At least one of send or recv must be non-null.
-	 *
-	 * @param send		Bytes to send to the device, or nullptr if
-	 *			no data is to be sent.
-	 * @param recv		Buffer for receiving bytes from the device,
-	 *			or nullptr if no bytes are to be received.
-	 * @param len		Number of bytes to transfer.
-	 * @return		OK if the exchange was successful, -errno
-	 *			otherwise.
-	 */
-	int		transfer(uint8_t *send, uint8_t *recv, unsigned len);
-
-	/**
-	 * Set the SPI bus frequency
-	 * This is used to change frequency on the fly. Some sensors
-	 * (such as the MPU6000) need a lower frequency for setup
-	 * registers and can handle higher frequency for sensor
-	 * value registers
-	 *
-	 * @param frequency	Frequency to set (Hz)
-	 */
-	void		set_frequency(uint32_t frequency);
-
-	/**
-	 * Set the SPI bus locking mode
-	 *
-	 * This set the SPI locking mode. For devices competing with NuttX SPI
-	 * drivers on a bus the right lock mode is LOCK_THREADS.
-	 *
-	 * @param mode	Locking mode
-	 */
-	void		set_lockmode(enum LockMode mode) { locking_mode = mode; }
-
-	LockMode	locking_mode;	/**< selected locking mode */
-
-private:
-	uint32_t			_device;
-	enum spi_mode_e		_mode;
-	uint32_t		_frequency;
-	struct spi_dev_s	*_dev;
-
-	/* this class does not allow copying */
-	SPI(const SPI &);
-	SPI operator=(const SPI &);
-
-protected:
-	int	_transfer(uint8_t *send, uint8_t *recv, unsigned len);
-
-	bool	external() { return px4_spi_bus_external(get_device_bus()); }
-
-};
-
-} // namespace device
-
-#endif /* _DEVICE_SPI_H */
+#ifdef __PX4_NUTTX
+#include "nuttx/SPI.hpp"
+#else
+#include "posix/SPI.hpp"
+#endif
